feat : add abilities for store(), update() and destroy() in ProductsController.php

+++ app/Http/Controllers/Api/ProductsController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\Api\Product\UpdateRequest;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Response;
 
 class ProductsController extends Controller
@@ -35,6 +36,12 @@ class ProductsController extends Controller
      */
     public function store(StoreRequest $request)
     {
+        $user = $request->user();
+        if(!$user->tokenCan('products.create')){
+            return response([
+                'message' => 'Not allowed'
+            ], 403);
+        }
         $product =  Product::create($request->validated());
 
         return Response::json($product, 201);
@@ -53,6 +60,12 @@ class ProductsController extends Controller
      */
     public function update(UpdateRequest $request, Product $product)
     {
+        $user = $request->user();
+        if(!$user->tokenCan('products.update')){
+            return response([
+                'message' => 'Not allowed'
+            ], 403);
+        }
         $product->update($request->validated());
 
         return Response::json($product);
@@ -63,6 +76,12 @@ class ProductsController extends Controller
      */
     public function destroy(string $id)
     {
+        $user = Auth::guard('sanctum')->user();
+        if(!$user->tokenCan('products.delete')){
+            return response([
+                'message' => 'Not allowed'
+            ], 403);
+        }
         Product::destroy($id);
         return response()->json(null, 204);
     }
